+++ packages/node/src/middleware/fastify.ts
@@ -120,7 +120,7 @@ const logTidePlugin: FastifyPluginCallback<FastifyMiddlewareOptions> = (
 };
 
 export const logTideFastifyPlugin = fp(logTidePlugin, {
-  fastify: '4.x',
+  fastify: '>=4.0.0',
   name: '@logtide/fastify-plugin',
 });